requirements.txt
asgiref==3.11.0
cffi==2.0.0
chardet==5.2.0
cryptography==46.0.3
Django==5.1.5
django-encrypted-model-fields==0.6.5
gunicorn==23.0.0
packaging==25.0
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.23
PyPDF2==3.0.1
python-decouple==3.8
reportlab==4.2.5
sqlparse==0.5.5
num2words==0.5.13
